+++ internal/config/constants.go
@@ -1,9 +1,10 @@
 package config
 
 const (
-	TAPA_ICON_NAME           string = "build/tapa.png"
+	TAPA_ICON_NAME           string = "build/appicon.png"
 	APP_NAME                 string = "TAPA"
 	APP_WIDTH                int    = 1024
+	APP_MIN_WIDTH            int    = 600
 	APP_HEIGHT               int    = 768
 	MACOS_ABOUT_MESSAGE      string = "The Actual Postman Alternative"
 	TAPA_DB_NAME             string = "tapa.db"